requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
coinbase-commerce==1.0.1
Django==4.1.5
django-bootstrap4==22.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
phonenumbers==8.13.4
Pillow==9.4.0
pycryptodome==3.16.0
pytz==2022.7.1
rave-python==1.2.16
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
